requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
azure-cognitiveservices-speech==1.42.0
azure-core==1.32.0
azure-identity==1.20.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==7.1.2
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.17.0
flatbuffers==1.12
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jiter==0.8.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
mutagen==1.47.0
namex==0.0.8
networkx==3.2.1
norbert==0.2.1
numba==0.60.0
numpy==1.23.5
oauthlib==3.2.2
openai==1.65.2
openai-whisper==20240930
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==1.5.3
portalocker==2.10.1
propcache==0.3.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pysubs2==1.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2025.1
pywin32==308
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.9.4
rsa==4.9
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
spleeter==2.4.0
sympy==1.13.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-cpu==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typer==0.3.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
whisper==1.1.10
wrapt==1.17.2
yarl==1.18.3
yt-dlp==2025.2.19
zipp==3.21.0
